requirements.txt
# requirements.txt

# Install the dependencies from this requirements.txt file by doing:
#   1: pip install --upgrade pip wheel
#   2: pip install -r requirements.txt

############################
# Installing Packages ######
############################
pip>=*
wheel>=*


############################
# General Packages #########
############################
polars~=0.20.21
requests>=*
loguru>=*
pyyaml>=*
gitpython>=*
orjson>=*
tqdm>=*
backoff>=*


############################
# Large Language Models ####
############################
ollama~=0.1.8
openai~=1.23.6


torch>=*
tensorflow>=*

###################
# Dev Tools #######
###################
pytest>=*
flake8>=*
black>=*
